Modification de la route DELETE pour prise en compte de l'authentification ainsi que de l'utilisation de fichiers.

+++ backend/controllers/stuff.js
@@ -1,4 +1,5 @@
 const Book = require('../models/book')
+const fs = require('fs')
 
 exports.createBook = (req, res, next) => {
     const bookObject = JSON.parse(req.body.book)
@@ -34,7 +35,24 @@ exports.getOneBook = (req, res, next) => {
 }
 
 exports.deleteOneBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Livre supprimé !' }))
-        .catch((error) => res.status(400).json({ error }))
+    Book.findOne({ _id: req.params.id })
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' })
+            } else {
+                const filename = book.imageUrl.split('/images/')[1]
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id })
+                        .then(() => {
+                            res.status(200).json({
+                                message: 'Livre supprimé !',
+                            })
+                        })
+                        .catch((error) => res.status(401).json({ error }))
+                })
+            }
+        })
+        .catch((error) => {
+            res.status(500).json({ error })
+        })
 }
